added some unit tests and pushed coverage to 10%

+++ AuctionHouse/backend/src/main/java/bg/dimitar/individual/business/impl/ItemManagerImpl.java
@@ -1,7 +1,7 @@
 package bg.dimitar.individual.business.impl;
 
 import bg.dimitar.individual.business.ItemManager;
-import bg.dimitar.individual.controller.dtos.Item;
+import bg.dimitar.individual.business.custom_exception.UnauthorizedChangeException;
 import bg.dimitar.individual.persistance.ItemRepository;
 import bg.dimitar.individual.persistance.entity.ItemEntity;
 import lombok.AllArgsConstructor;
@@ -17,18 +17,14 @@ public class ItemManagerImpl implements ItemManager {
     private ItemRepository repository;
 
     @Override
-    public ItemEntity getItemByID(long id) {
+    public ItemEntity getItemByID(Long id) {
         final Optional<ItemEntity> itemOptional = repository.findById(id);
         return itemOptional.orElse(null);
     }
 
     @Override
     public List<ItemEntity> getAllItems() {
-        List<ItemEntity> returnItems = new ArrayList<>();
-
-        returnItems.addAll(repository.findAll());
-
-        return returnItems;
+        return new ArrayList<>(repository.findAll());
     }
 
     @Override
@@ -36,13 +32,27 @@ public class ItemManagerImpl implements ItemManager {
         return repository.save(item);
     }
 
-//    @Override
-//    public boolean updateItem(ItemEntity item) {
-//        return repository.updateItem(item);
-//    }
+    @Override
+    public boolean updateItem(ItemEntity item, Long userId) throws UnauthorizedChangeException {
+        Optional<ItemEntity> itemFromDb = repository.findById(item.getId());
+
+        if (itemFromDb.isPresent() && !itemFromDb.get().getPostedByUserId().equals(userId)) {
+            throw new UnauthorizedChangeException("You do not have permission to edit this item");
+        }
+
+        repository.save(item);
+        return true;
+    }
 
     @Override
-    public void deleteItem(long id) {
+    public boolean deleteItem(Long id, Long userId) throws UnauthorizedChangeException {
+        Optional<ItemEntity> itemFromDb = repository.findById(id);
+
+        if (itemFromDb.isPresent() && !itemFromDb.get().getPostedByUserId().equals(userId)) {
+            throw new UnauthorizedChangeException("You do not have permission to delete this item");
+        }
+
         repository.deleteById(id);
+        return true;
     }
 }
